refactor: 拆分设置页个人安全配置
